requirements.txt
streamlit>=1.35.0
transformers>=4.40.0
torch>=2.2.0
Pillow>=10.0.0
ftfy>=6.1.1
regex>=2023.12.25
numpy>=1.26.0
anthropic>=0.28.0
